pyproject.toml
[project]
name = "runpod-tts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.2.5",
    "orpheus-speech>=0.1.0",
    "pynput>=1.8.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "runpod>=1.7.9",
    "scipy>=1.15.2",
    "sounddevice>=0.5.1",
]
